+++ aoc2019/src/main/kotlin/intcode/Utils.kt
@@ -1,3 +1,3 @@
 package intcode
 
-fun parseIntcodeProgram(raw: String) = raw.replace("\n", "").trim().split(",").map { it.toInt() }
+fun parseIntcodeProgram(raw: String) = raw.replace("\n", "").trim().split(",").map { it.toLong() }
